Do not print generated password by default
Fixes #410
